Creació Repositoris MongoDB i comprovació inserir dades.

See merge request mayo.ortiz.adria/projecte2_openroad!14

+++ src/main/java/com/copernic/projecte2_openroad/model/mysql/Incidencia.java
@@ -47,7 +47,7 @@ public class Incidencia {
     @Column(nullable = false)
     private String motiu;
     @Column(nullable = false)
-    private Float cost;
+    private Double cost;
     @Column(nullable = false, name = "data_inici")
     private LocalDate dataInici;
